Merged PR 54: Capability Assessment w/ Data Model Integration.
This PR contains the implementation of the Capability Assessment Pages.

As everything was implemented and then had to be rebased against a bunch of changes to the API, there are likely things I missed that were implemented work around some of the features that were missing from the Data Model.