Revert "Merged PR 102: SolutionExByOrganisation
An endpoint for requesting all SolutionEx information stored against an organization has been added. This means that a single call can be issued from the front end, and a single called from the back-end can be issued to the CRM, drastically improving the performance as more solutions are added to an organization.

There seems to be a bunch of seed data changes made in a previous commit, but the handling of those changes don't seem to have come with that commit. Thus the regression tests are currently failing, however, the Solution Dashboard tests did pass.

The changes in this PR can be summarised as:
1. Regeneration of Swagger Endpoints.
2. Update of `solutionsForSupplierDashboard`"