☣️ #987 EXPERIMENTAL Auto-restore field inputs if redirected due to authentication issue
Experimental and rather inelegant but effective as a proof-of-concept.

If an authentication failure occurs AND there is something being posted in the request body,
encode the posted data in the redirect URL to be picked up later by the code in app.js which
restores it and updates the character counts.